+++ build_mainline.sh
@@ -1,6 +1,6 @@
 #!/bin/sh -e
 #
-# Copyright (c) 2009-2015 Robert Nelson <robertcnelson@gmail.com>
+# Copyright (c) 2009-2016 Robert Nelson <robertcnelson@gmail.com>
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -33,29 +33,36 @@ patch_kernel () {
 
 	if [ ! "${RUN_BISECT}" ] ; then
 		git add --all
-		git commit --allow-empty -a -m "${KERNEL_TAG}-${BUILD} patchset"
+		git commit --allow-empty -a -m "${KERNEL_TAG}${BUILD} patchset"
 	fi
 
-	cd ${DIR}/
+	cd "${DIR}/" || exit
 }
 
 copy_defconfig () {
-	cd ${DIR}/KERNEL/
-	make ARCH=arm CROSS_COMPILE="${CC}" distclean
-	make ARCH=arm CROSS_COMPILE="${CC}" ${config}
-	cp -v .config ${DIR}/patches/HEAD_${config}
-	cd ${DIR}/
+	cd "${DIR}/KERNEL" || exit
+	make ARCH=${KERNEL_ARCH} CROSS_COMPILE="${CC}" distclean
+	make ARCH=${KERNEL_ARCH} CROSS_COMPILE="${CC}" ${config}
+	cp -v .config "${DIR}/patches/HEAD_${config}"
+	cd "${DIR}/" || exit
 }
 
 make_menuconfig () {
 	cd "${DIR}/KERNEL" || exit
-	make ARCH=arm CROSS_COMPILE="${CC}" menuconfig
-	cp -v .config "${DIR}/patches/defconfig"
+	make ARCH=${KERNEL_ARCH} CROSS_COMPILE="${CC}" menuconfig
+	if [ ! -f "${DIR}/.yakbuild" ] ; then
+		cp -v .config "${DIR}/patches/defconfig"
+	fi
 	cd "${DIR}/" || exit
 }
 
 make_kernel () {
-	image="zImage"
+	if [ "x${KERNEL_ARCH}" = "xarm" ] ; then
+		image="zImage"
+	else
+		image="Image"
+	fi
+
 	unset address
 
 	##uImage, if you really really want a uImage, zreladdr needs to be defined on the build line going forward...
@@ -65,15 +72,15 @@ make_kernel () {
 
 	cd "${DIR}/KERNEL" || exit
 	echo "-----------------------------"
-	echo "make -j${CORES} ARCH=arm LOCALVERSION=-${BUILD} CROSS_COMPILE=\"${CC}\" ${address} ${image} modules"
+	echo "make -j${CORES} ARCH=${KERNEL_ARCH} LOCALVERSION=${BUILD} CROSS_COMPILE=\"${CC}\" ${address} ${image} modules"
 	echo "-----------------------------"
-	make -j"${CORES}" ARCH=arm LOCALVERSION=-"${BUILD}" CROSS_COMPILE="${CC}" "${address}" "${image}" modules
+	make -j${CORES} ARCH=${KERNEL_ARCH} LOCALVERSION=${BUILD} CROSS_COMPILE="${CC}" ${address} ${image} modules
 	echo "-----------------------------"
 
-	if grep -q dtbs "${DIR}/KERNEL/arch/arm/Makefile"; then
-		echo "make -j${CORES} ARCH=arm LOCALVERSION=-${BUILD} CROSS_COMPILE=\"${CC}\" dtbs"
+	if grep -q dtbs "${DIR}/KERNEL/arch/${KERNEL_ARCH}/Makefile"; then
+		echo "make -j${CORES} ARCH=${KERNEL_ARCH} LOCALVERSION=${BUILD} CROSS_COMPILE=\"${CC}\" dtbs"
 		echo "-----------------------------"
-		make -j"${CORES}" ARCH=arm LOCALVERSION=-"${BUILD}" CROSS_COMPILE="${CC}" dtbs
+		make -j${CORES} ARCH=${KERNEL_ARCH} LOCALVERSION=${BUILD} CROSS_COMPILE="${CC}" dtbs
 		echo "-----------------------------"
 	fi
 
@@ -84,8 +91,8 @@ make_kernel () {
 		rm -rf "${DIR}/deploy/config-${KERNEL_UTS}" || true
 	fi
 
-	if [ -f ./arch/arm/boot/${image} ] ; then
-		cp -v arch/arm/boot/${image} "${DIR}/deploy/${KERNEL_UTS}.${image}"
+	if [ -f ./arch/${KERNEL_ARCH}/boot/${image} ] ; then
+		cp -v arch/${KERNEL_ARCH}/boot/${image} "${DIR}/deploy/${KERNEL_UTS}.${image}"
 		cp -v .config "${DIR}/deploy/config-${KERNEL_UTS}"
 	fi
 
@@ -112,30 +119,30 @@ make_pkg () {
 	if [ -d "${DIR}/deploy/tmp" ] ; then
 		rm -rf "${DIR}/deploy/tmp" || true
 	fi
-	mkdir -p ${DIR}/deploy/tmp
+	mkdir -p "${DIR}/deploy/tmp"
 
 	echo "-----------------------------"
 	echo "Building ${pkg} archive..."
 
 	case "${pkg}" in
 	modules)
-		make -s ARCH=arm CROSS_COMPILE="${CC}" modules_install INSTALL_MOD_PATH="${DIR}/deploy/tmp"
+		make -s ARCH=${KERNEL_ARCH} LOCALVERSION=${BUILD} CROSS_COMPILE="${CC}" modules_install INSTALL_MOD_PATH="${DIR}/deploy/tmp"
 		;;
 	firmware)
-		make -s ARCH=arm CROSS_COMPILE="${CC}" firmware_install INSTALL_FW_PATH="${DIR}/deploy/tmp"
+		make -s ARCH=${KERNEL_ARCH} LOCALVERSION=${BUILD} CROSS_COMPILE="${CC}" firmware_install INSTALL_FW_PATH="${DIR}/deploy/tmp"
 		;;
 	dtbs)
-		if grep -q dtbs_install "${DIR}/KERNEL/arch/arm/Makefile"; then
-			make -s ARCH=arm LOCALVERSION=-"${BUILD}" CROSS_COMPILE="${CC}" dtbs_install INSTALL_DTBS_PATH="${DIR}/deploy/tmp"
+		if grep -q dtbs_install "${DIR}/KERNEL/arch/${KERNEL_ARCH}/Makefile"; then
+			make -s ARCH=${KERNEL_ARCH} LOCALVERSION=${BUILD} CROSS_COMPILE="${CC}" dtbs_install INSTALL_DTBS_PATH="${DIR}/deploy/tmp"
 		else
-			find ./arch/arm/boot/ -iname "*.dtb" -exec cp -v '{}' "${DIR}/deploy/tmp/" \;
+			find ./arch/${KERNEL_ARCH}/boot/ -iname "*.dtb" -exec cp -v '{}' "${DIR}/deploy/tmp/" \;
 		fi
 		;;
 	esac
 
 	echo "Compressing ${KERNEL_UTS}${deployfile}..."
 	cd "${DIR}/deploy/tmp" || true
-	tar "${tar_options}" "../${KERNEL_UTS}${deployfile}" ./*
+	tar ${tar_options} "../${KERNEL_UTS}${deployfile}" ./*
 
 	cd "${DIR}/" || exit
 	rm -rf "${DIR}/deploy/tmp" || true
@@ -162,6 +169,14 @@ make_dtbs_pkg () {
 	pkg="dtbs"
 	make_pkg
 }
+
+if [  -f "${DIR}/.yakbuild" ] ; then
+	if [ -f "${DIR}/recipe.sh.sample" ] ; then
+		if [ ! -f "${DIR}/recipe.sh" ] ; then
+			cp -v "${DIR}/recipe.sh.sample" "${DIR}/recipe.sh"
+		fi
+	fi
+fi
 
 /bin/sh -e "${DIR}/tools/host_det.sh" || { exit 1 ; }
 
@@ -193,6 +208,9 @@ fi
 unset CC
 unset LINUX_GIT
 . "${DIR}/system.sh"
+if [  -f "${DIR}/.yakbuild" ] ; then
+	. "${DIR}/recipe.sh"
+fi
 /bin/sh -e "${DIR}/scripts/gcc.sh" || { exit 1 ; }
 . "${DIR}/.CC"
 echo "CROSS_COMPILE=${CC}"
@@ -214,16 +232,21 @@ if [ "${FULL_REBUILD}" ] ; then
 		/bin/sh -e "${DIR}/scripts/bisect.sh" || { exit 1 ; }
 	fi
 
-	#patch_kernel
+#	if [ ! -f "${DIR}/.yakbuild" ] ; then
+#		patch_kernel
+#	fi
 	copy_defconfig
 fi
 if [ ! "${AUTO_BUILD}" ] ; then
 	make_menuconfig
 fi
+if [  -f "${DIR}/.yakbuild" ] ; then
+	BUILD=$(echo ${kernel_tag} | sed 's/[^-]*//'|| true)
+fi
 make_kernel
 make_modules_pkg
 make_firmware_pkg
-if grep -q dtbs "${DIR}/KERNEL/arch/arm/Makefile"; then
+if grep -q dtbs "${DIR}/KERNEL/arch/${KERNEL_ARCH}/Makefile"; then
 	make_dtbs_pkg
 fi
 echo "-----------------------------"
